Интеграция фронта и бека

+++ back/src/main/java/org/reactivetales/blog/service/ArticleService.java
@@ -1,9 +1,9 @@
-package org.example.blog.service;
+package org.reactivetales.blog.service;
 
 import lombok.RequiredArgsConstructor;
-import org.example.blog.dto.CreateArticleRequest;
-import org.example.blog.persistence.entity.Article;
-import org.example.blog.persistence.repository.ArticleRepository;
+import org.reactivetales.blog.persistence.dto.CreateArticleRequest;
+import org.reactivetales.blog.persistence.entity.Article;
+import org.reactivetales.blog.repository.ArticleRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
@@ -12,6 +12,7 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class ArticleService {
+
     private final ArticleRepository articleRepository;
 
     public List<Article> getAllArticles() {
@@ -20,15 +21,15 @@ public class ArticleService {
 
     public Article getArticleById(String id) {
         return articleRepository.findById(id)
-                .orElseThrow(() -> new IllegalStateException("Сущность не существует!"));
+                .orElseThrow(() -> new IllegalStateException("Article not found!"));
     }
 
     public Article createOrUpdate(CreateArticleRequest createArticleRequest) {
-        Article article = StringUtils.hasText(createArticleRequest.getId()) ?
-                articleRepository.getById(createArticleRequest.getId())
+        Article article = StringUtils.hasText(createArticleRequest.getId())
+                ? articleRepository.getById(createArticleRequest.getId())
                 : new Article();
 
-        article.setHeader(createArticleRequest.getTitle());
+        article.setHeader(createArticleRequest.getHeader());
         article.setContent(createArticleRequest.getContent());
         articleRepository.save(article);
 
@@ -36,6 +37,10 @@ public class ArticleService {
     }
 
     public void delete(String id) {
+        if (!articleRepository.existsById(id)) {
+            throw new IllegalStateException("Article not found!");
+        }
+
         articleRepository.deleteById(id);
     }
 }
